requirements.txt
torch
numpy>=1.15.2
opencv-python
torchvision
pillow>=8.1.1
scikit-image
scikit-learn
paddlepaddle
paddlehub
gradio